+++ NEWS.md
@@ -9,4 +9,5 @@
 * Support FSSP type of Secure Store (APPPOCTOOL-59)
 * Refresh entitlement info upon failed entitlement check (MODSIDECAR-126)
 * GET Requests With Bodies Are Not Forwarded Properly (MODSIDECAR-125)
+* Consume "System User required" attribute from module bootstrap and request system user token only if the attribute value is true (MODSIDECAR-87)
 * Remove HTTP priority header to work around DoS CVE-2025-31650 ([MODSIDECAR-130](https://folio-org.atlassian.net/browse/MODSIDECAR-130))